impleemnt common blocking_gates_warning on prologue and lesson pages
